requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
packaging==23.2
pip==22.3.1
prompt-toolkit==3.0.39
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.2
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
gunicorn
Pillow
flower